pyproject.toml
[tool.poetry]
name = "rpi-electrical_panel"
version = "1.0.0"
description = ""
authors = ["Nicolas HERNANDEZ <nicolas.hernandezpaez@orange.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "2.0.3"
flask-smorest = "^0.34.1"
PyYAML = "^5.4.1"
paho-mqtt = "^1.6.1"
bson = "^0.5.10"
#smbus = "^1.1.post2"
timeloop = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-mock = "^3.8.1"
black = "^22.3.0"
pylint = "^2.14.3"
pytest-cov = "^3.0.0"
